Move Verdict to its own module since it does not depend only on 'immediate' expressions
